Add files via upload
Names Online is a front end module for Joomla 3.10 and 4.0.  It displays the number of visitors (Guests) and members (Registered users and above), who are logged in and  currently accessing the website, with the full name of those users.

The Options of this module allow you to display only the numbers of guests and members, or the registered full names of the logged-in users, or both.  If the option Filter Groups is selected, the users can see only the names of users in the user's own own user group and its child groups.

The initial version of this module was cloned from the Who's Online module, mod_whosonline, of Joomla 3.10 and tested in Joomla 3.10.4 and Joomla 4.0.5 using PHP 7.4.26, 8.0.13 and 8.1.0.